parallelize: only check to-be-rewritten commits for rewritability
closes #6587

Given a commit history like this:
    ○  ccc
    ○  bbb
    ◆  aaa

Parallelize *should* be able to create this:
    ○  ccc
    │ ○  bbb
    ├─╯
    │ ◆  aaa
    ├─╯

This is fine because `aaa` will not be rewritten --- its children will,
but neither `aaa`'s parents nor author information needs to be changed.

However, the current implementation checks _all_ commits in the
target set for rewritability, instead of only the commits that will
be rewritten.

This commit changes that. Now, only the commits whose parents *will*
change are checked for rewritability.