index: use bit set to omit visited ancestors in is_ancestor_pos()
This seems good for both lucky and unlucky cases. If the ancestor is reachable,
DFS tends to finish within a fewer steps. If the ancestor is unreachable, we'll
have to visit all candidates, so the visited set can be large.

jj bench is-ancestor --ignore-working-copy -R ~/mirrors/linux
```
group                       old                     new
-----                       ---                     ---
is-ancestor-v6.0-v6.10      2.45     93.0±0.46µs    1.00     37.9±0.19µs
is-ancestor-v6.0.1-v6.10    2.90     12.7±0.09ms    1.00      4.4±0.13ms
```