async: add async version of get_root_tree()
This introduces an async version of `get_root_tree()` which reads
trees that are part of a conflict concurrently. I haven't noticed a
need for that at Google or elsewhere but it could be significant.

I updated the only caller I could find that was already in an async
context.