fix: write complete stderr output at once from fix tools
Fix tools often emit information on stderr for the user to see. Since
fix tools are run in parallel and stderr is unbuffered, this information
can be jumbled together. This seems to happen especially frequently on
Windows.

This change locks stderr before writing to it so the stderr output from
a tool is written all at once.