merged_tree: poll trees in the order we're going to emit them
It should genenerally be better to prioritize polling trees in the
order we're going to emit their entries. For example, if we have
pending trees `zzz/` and `dir/aaa/`, it's better to poll the latter
even though we inserted the former first.

This also prepares for fixing a bug related to the order we emit. We
will then want to look up in `pending_trees` by key found in `items`.