clippy: run autofix to use `Self` more often
For some reason, running `cargo update` makes clippy notice these can
be replaced with `Self`. I put this commit first, though, since it is
helpful regardless of the update.

This is the result of the following command after the `cargo update`
from the next commit (even though the refactor is valid regardless of
the `cargo update`).

```
cargo +nightly clippy  --workspace --all-targets --all-features --fix
```