docs: provide analogue for git checkout
This commit includes a jj recipe for the Git command table that is
approximately equivalent to Git's checkout command, which can be used to
examine source code at named revision or branch.  This is indispensible
for learn-by-doing types who are curious how to remap Git usage back
onto jj with minimal effort.