nextest: change the "SLOW" timeout in CI to 10s
Previously, nextest marked tests slow in CI when they run for more
than 5 seconds. This makes it so that this happens after 10 seconds.

I did *not* change the timeout for when tests are actually *terminated*
for being too slow. It remains at 100s. It might make sense to increase
it (keeping the `terminate-after` multiplier at 10).

This is just adjusting to taste, but I find the number of SLOWs in
https://github.com/jj-vcs/jj/actions/runs/17843955040/job/50747342340?pr=7549
distracting, and most of the slow tests on Windows last more than 10s
anyway.

We could also do this only on Windows if people prefer, but
I think doing it everywhere is simpler.