docs: add note about PowerShell `$PROFILE` execution policy
On Windows, PowerShell scripts are not allowed to be executed by default
due to the default execution policy `Restricted`. This includes the
`$PROFILE` script (even though PowerShell will still happily attempt to
source it if it exist, resulting in "running scripts is disabled on this
system" errors every time it starts... go figure...).

To avoid this, you generally need to set the execution policy to at
least `RemoteSigned` (which will only block scripts downloaded from the
internet which are not signed; this is the default on Windows Server).
This can be done by running

    Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope CurrentUser

as an administrator, though there tends to be [discussions][1] about the
security tradeoffs of scope and granularity, so I think it is best to
link to the MS documentation on execution policies to allow users to
draw their own conclusions. 

This issue does not affect PowerShell users on Linux or macOS.

[1]: https://superuser.com/questions/106360/how-to-enable-execution-of-powershell-scripts