cli: complete: Filter out bad lines
For completion of workspace names, we are parsing a content where each
line should have a tab, separating a candidate workspace name from the
working copy commit description, serving as help text for the value.

If we cannot parse one of the lines, instead of proposing a bad value,
print an error message and ignore the line.